d&d Ref 0.5.1
Added villanous class options from the DMG, with a disclaimer

Changed colors to illustrate if a class option came from the player's handbook, XGE, or the DMG

Began adding races. They're all in there, but subraces are in their own card. This has been remedied for Aasimar, Dwarves, and Warforged so far.